requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export
#-e .
a2wsgi==1.10.7
annotated-types==0.7.0
anyio==4.7.0 
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
connexion==3.1.0
distlib==0.3.9
dnspython==2.7.0
filelock==3.16.1
flask==3.1.0
h11==0.14.0
hatchling==1.26.3
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markupsafe==3.0.2
mongomock==4.3.0
mypy==1.14.1
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.10.3
pydantic-core==2.27.1
pymongo==4.10.1
pyproject-api==1.8.0
pytest==8.3.4
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
pyyaml==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
ruff==0.8.2 
sentinels==1.0.0
sniffio==1.3.1
starlette==0.41.3
swagger-ui-bundle==1.1.0
tox==4.23.2 
trove-classifiers==2024.10.21.16
typing-extensions==4.12.2
urllib3==2.2.3 
uvicorn==0.32.1 
uvloop==0.21.0
virtualenv==20.28.1
watchfiles==1.0.0
websockets==14.1
werkzeug==3.1.3
python-jose[cryptography]==3.5.0